# src/swagger.yaml
openapi: 3.0.0
info:
  title: OA-Pank API
  version: 1.0.1
  description: API documentation for OA-Pank banking system with RFC 7231 compliant responses
servers:
  - url: /
    description: Current server
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        error:
          type: string
          example: error_code
        details:
          type: string
          example: Detailed error information
...
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: username
              message:
                type: string
                example: Username is required
              value:
                type: string
                example: ""
...
    Account:
      type: object
      properties:
        _id:
          type: string
          example: "60d21b4667d0d8992e610c85"
        accountNumber:
          type: string
          example: OAP12345678
        user:
          type: string
          example: 60d21b4667d0d8992e610c85
        balance:
          type: number
          format: float
          example: 1000.00
        currency:
          type: string
          enum: [EUR, USD, GBP, CHF, JPY, AUD, CAD, SEK, NOK, DKK, PLN, CZK]
          example: EUR
        isActive:
          type: boolean
          example: true
        type:
          type: string
          enum: [checking, savings, investment]
          example: checking
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AccountInput:
      type: object
      required:
        - currency
        - type
      properties:
        currency:
          type: string
          enum: [EUR, USD, GBP, CHF, JPY, AUD, CAD, SEK, NOK, DKK, PLN, CZK]
          example: EUR
        type:
          type: string
          enum: [checking, savings, investment]
          example: checking
...
paths:
...
  /accounts:
    get:
      tags:
        - Accounts
      summary: Get all accounts
      description: Returns all accounts for the authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Accounts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  accounts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Account"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Accounts
      summary: Create account
      description: Creates a new account for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  account:
                    $ref: "#/components/schemas/Account"
        "400":
          description: Bad request - Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /accounts/{id}:
    get:
      tags:
        - Accounts
      summary: Get account by ID
      description: Returns a specific account's information
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Account ID
      responses:
        "200":
          description: Account information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  account:
                    $ref: "#/components/schemas/Account"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - No permission to access this account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found - Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - Accounts
      summary: Delete account
      description: Deletes an account
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Account ID
      responses:
        "200":
          description: Account deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Account deleted successfully"
        "409":
          description: Conflict - Account has non-zero balance
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    format: uri
                    example: "https://example.com/conflict"
                  title:
                    type: string
                    example: "Resource Conflict"
                  status:
                    type: integer
                    example: 409
                  detail:
                    type: string
                    example: "The account has a non-zero balance which conflicts with the deletion request"
                  instance:
                    type: string
                    format: uri
                    example: "/accounts/123"
                  accountId:
                    type: string
                    example: "123"
                  balance:
                    type: number
                    example: 100.50
                  currency:
                    type: string
                    example: "EUR"
                  success:
                    type: boolean
                    example: false
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - No permission to delete this account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found - Account not found
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    format: uri
                    example: "https://example.com/not-found"
                  title:
                    type: string
                    example: "Resource Not Found"
                  status:
                    type: integer
                    example: 404
                  detail:
                    type: string
                    example: "No account found with the provided ID"
                  instance:
                    type: string
                    format: uri
                    example: "/accounts/123"
                  accountId:
                    type: string
                    example: "123"
                  success:
                    type: boolean
                    example: false
...